Move Lock into core-kernel utils
